int price

+++ resources/views/component/cart_table.blade.php
@@ -20,13 +20,13 @@
                             <p class="text-center">{{$product->product->name}}</p>
                         </a>
                     </td>
-                    <td class="text-center padding-top-60">{{$product->product->price}} грн</td>
+                    <td class="text-center padding-top-60">{{(int)$product->product->price}} грн</td>
                     <td class="text-center"><!-- Quinty -->
 
                         <div class="quinty padding-top-20">
                             <input id="count{{$product->product->id}}" type="number" value="{{$product->count}}" oninput="changeCount({{$product->product->id}},{{$product->cart_id}},'{{route('product_count')}}')">
                         </div></td>
-                    <td class="text-center padding-top-60" id="price{{$product->product->id}}">{{((double)$product->product->price * (integer)$product->count)}} грн</td>
+                    <td class="text-center padding-top-60" id="price{{$product->product->id}}">{{((int)$product->product->price * (integer)$product->count)}} грн</td>
                     <td class="text-center padding-top-60"><a href="#." class="remove" onclick="deleteProduct({{$product->product->id}},{{$product->cart_id}},'{{route('product_delete')}}'); return false;"><i class="fa fa-close"></i></a></td>
                 </tr>
             @empty
@@ -55,12 +55,12 @@
                                 <p>{{$product->product->short_description}}</p>
                             </div>
                         </div>
-                        <div>{{$product->product->price}} грн <span class="show-480">{{__('  - за 1 единицу')}}</span></div>
+                        <div>{{(int)$product->product->price}} грн <span class="show-480">{{__('  - за 1 единицу')}}</span></div>
                         <div class="quinty">
                             <input id="count-mob{{$product->product->id}}" type="number" value="{{$product->count}}" oninput="changeCount({{$product->product->id}},{{$product->cart_id}},'{{route('product_count')}}')">
                         </div>
                         <div id="price-mob{{$product->product->id}}">
-                            {{((double)$product->product->price * (integer)$product->count)}} грн <span class="show-480">{{__(' - общяя стоимость')}}</span>
+                            {{((int)$product->product->price * (integer)$product->count)}} грн <span class="show-480">{{__(' - общяя стоимость')}}</span>
                         </div>
                         <div>
                             <a href="#." class="remove" onclick="deleteProduct({{$product->product->id}},{{$product->cart_id}},'{{route('product_delete')}}'); return false;"><i class="fa fa-close"></i></a>
@@ -88,7 +88,7 @@
                             }
                         }
                 @endphp
-        {{$sum}} грн
+        {{(int)$sum}} грн
             </span>
 </h6>
